# contracts/generated/openapi/pollen.v1.yaml
openapi: 3.1.0
info:
  title: Pollen DEK Control-Plane API
  version: 1.0.0
  contact:
    name: Pollen Platform Team
    url: https://your-cloud-domain.com
    email: platform@your-cloud-domain.com
  description: Contract between Pollen Cloud / Local Control Plane and the DEK edge enforcement plane.
...
paths:
...
  /v1/a2a/message:
    post:
      operationId: A2AApi_sendMessage
      description: Send an Inter-Agent Trust Protocol (IATP) message via the DEK mediator.
      parameters:
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/A2AMessageResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - A2A Mediator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/A2AMessage'
...
components:
  parameters:
    PollenHeaders.contractVersion:
      name: X-Pollen-Contract-Version
      in: header
      required: true
      schema:
        type: string
    PollenHeaders.deviceId:
      name: X-Pollen-Device-Id
      in: header
      required: false
      schema:
        type: string
    PollenHeaders.tenantId:
      name: X-Pollen-Tenant-Id
      in: header
      required: false
      schema:
        type: string
  schemas:
    A2AMessage:
      type: object
      required:
        - schema_version
        - sender_id
        - receiver_id
        - payload_encrypted
        - signature
      properties:
        schema_version:
          type: string
          enum:
            - a2a-message.v1
        sender_id:
          type: string
        receiver_id:
          type: string
        payload_encrypted:
          type: string
        signature:
          type: string
    A2AMessageResponse:
      type: object
      required:
        - schema_version
        - status
      properties:
        schema_version:
          type: string
          enum:
            - a2a-response.v1
        status:
          type: string
        reason:
          type: string
...
    PollenError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string
...
servers:
  - url: http://127.0.0.1:17891
    description: Local Control Plane / Mock Cloud
    variables: {}
  - url: https://api.pollen.cloud
    description: Pollen Cloud
    variables: {}