# crates/dek-core/src/api.rs
// SPDX-License-Identifier: Apache-2.0
// Copyright (c) 2026 AEC Infraconnect

use arc_swap::ArcSwap;
use axum::{
    extract::State,
    http::StatusCode,
    response::IntoResponse,
    routing::{get, post},
    Json, Router,
};
use dek_activation::snapshot::RuntimeSnapshot;
use dek_decision::{DecisionRequest, DecisionResponse};
use std::sync::Arc;
use tracing::{error, info};

use dek_policy_syncer::EnforcementState;
use dek_telemetry::CloudTelemetrySink;

#[derive(Clone)]
struct ApiState {
    snapshot: Arc<ArcSwap<RuntimeSnapshot>>,
    enforcement: Arc<ArcSwap<EnforcementState>>,
    telemetry: Option<Arc<CloudTelemetrySink>>,
    identity_health: tokio::sync::watch::Receiver<crate::svid_renewal_failclosed::IdentityHealth>,
}

pub async fn start_sidecar_api(
    snapshot: Arc<ArcSwap<RuntimeSnapshot>>,
    enforcement: Arc<ArcSwap<EnforcementState>>,
    telemetry: Option<Arc<CloudTelemetrySink>>,
    identity_health: tokio::sync::watch::Receiver<crate::svid_renewal_failclosed::IdentityHealth>,
    port: u16,
) -> anyhow::Result<()> {
    let state = ApiState {
        snapshot,
        enforcement,
        telemetry,
        identity_health,
    };

    let app = Router::new()
        .route("/v1/healthz", get(healthz))
        .route("/v1/readyz", get(readyz))
        .route("/v1/capabilities", get(capabilities))
        .route("/v1/decision/check", post(check))
        .route("/v1/decision/batch-check", post(batch_check))
        .route("/v1/a2a/message", post(a2a_message))
        .with_state(state);

    let addr = format!("127.0.0.1:{}", port);
    info!("Starting Sidecar API on {}", addr);
    let socket_addr: std::net::SocketAddr = addr.parse()?;
    let socket = if socket_addr.is_ipv6() {
        tokio::net::TcpSocket::new_v6()?
    } else {
        tokio::net::TcpSocket::new_v4()?
    };
    let _ = socket.set_reuseaddr(true);
    socket.bind(socket_addr)?;
    let listener = socket.listen(1024)?;

    tokio::spawn(async move {
        if let Err(e) = axum::serve(listener, app).await {
            error!("Sidecar API server failed: {}", e);
        }
    });

    Ok(())
}

async fn healthz() -> impl IntoResponse {
    (StatusCode::OK, "OK")
}

async fn readyz() -> impl IntoResponse {
    (StatusCode::OK, "READY")
}

async fn capabilities() -> impl IntoResponse {
    Json(dek_domain_schema::EnforcementCapabilities::detect())
}

async fn check(
    State(state): State<ApiState>,
    Json(req): Json<DecisionRequest>,
) -> impl IntoResponse {
    let enf = state.enforcement.load();
    let health = *state.identity_health.borrow();

    let is_strict_deny = matches!(**enf, EnforcementState::StrictDeny { .. })
        || health == crate::svid_renewal_failclosed::IdentityHealth::Expired;

    if is_strict_deny {
        let reason = if health == crate::svid_renewal_failclosed::IdentityHealth::Expired {
            "SVID expired".to_string()
        } else if let EnforcementState::StrictDeny { ref reason, .. } = **enf {
            reason.clone()
        } else {
            "Unknown".to_string()
        };

        let response = DecisionResponse {
            decision_id: uuid::Uuid::new_v4().to_string(),
            allow: false,
            reason_code: "DENIED_BY_FAIL_CLOSED".into(),
            reason: format!("fail-closed: {}", reason),
            obligations: vec![],
            effects: serde_json::json!({}),
            policy_bundle_id: "active".into(),
            policy_bundle_version: "v1".into(),
            evaluator_results: vec![],
            latency_ms: 0_i64,
        };
        return (
            StatusCode::OK,
            Json(serde_json::to_value(response).unwrap_or_else(|_| serde_json::json!({}))),
        );
    }

    let snap = state.snapshot.load();
    let val = serde_json::to_value(&req).unwrap_or(serde_json::json!({}));

    let start = std::time::Instant::now();
    let res =
        snap.router
            .authorize(val)
            .await
            .unwrap_or_else(|_| dek_policy_runtime::PolicyDecision {
                evaluator_id: "core_api".into(),
                evaluator_type: "router".into(),
                required: true,
                status: "error".into(),
                decision: "deny".into(),
                allow: false,
                reason: "Policy evaluation failed".into(),
                effects: serde_json::json!({}),
                obligations: vec![],
                metadata: serde_json::json!({}),
                explanation: None,
                user_action_required: false,
                user_action_th: None,
            });

    let latency = start.elapsed().as_millis() as i64;

    let has_require_approval = res.obligations.iter().any(|o| o == "require_approval");
    let allow = res.allow && !has_require_approval;

    let response = DecisionResponse {
        decision_id: uuid::Uuid::new_v4().to_string(),
        allow,
        reason_code: if has_require_approval {
            "PENDING_APPROVAL".into()
        } else if allow {
            "OK".into()
        } else {
            "DENIED_BY_POLICY".into()
        },
        reason: res.reason.clone(),
        obligations: res
            .obligations
            .into_iter()
            .map(|o| dek_decision::Obligation {
                kind: o,
                parameters: serde_json::json!({}),
            })
            .collect(),
        effects: res.effects,
        policy_bundle_id: "active".into(),
        policy_bundle_version: "v1".into(),
        evaluator_results: vec![],
        latency_ms: latency,
    };

    let mut json_resp = serde_json::to_value(&response).unwrap_or_else(|_| serde_json::json!({}));
    if has_require_approval {
        if let Some(obj) = json_resp.as_object_mut() {
            obj.insert(
                "error".to_string(),
                serde_json::json!({
                    "code": -32002,
                    "message": "Access Denied: pending_approval",
                    "data": {
                        "status": "denied",
                        "reason": res.reason
                    }
                }),
            );
        }
    }

    if let Some(telemetry) = &state.telemetry {
        let event = serde_json::json!({
            "schema_version": "1.0",
            "event_id": response.decision_id,
            "event_type": "decision",
            "timestamp": chrono::Utc::now().to_rfc3339(),
            "tenant_id": req.context.get("tenant_id").and_then(|v| v.as_str()).unwrap_or("local").to_string(),
            "device_id": "api-sidecar",
            "trace_id": "",
            "span_id": "",
            "spiffe_id": "",
            "pep_type": "api",
            "agent_id": "dek",
            "principal_id": req.principal,
            "mcp_server_id": "",
            "tool_id": "",
            "tool_name": "",
            "action": req.action,
            "resource_id": req.resource,
            "resource_uri": "",
            "decision": if res.allow { "allow" } else { "deny" },
            "reason": res.reason,
            "policy_ids": Vec::<String>::new(),
            "bundle_id": "active",
            "bundle_version": "v1",
            "latency_ms": latency,
            "cached": false,
            "redaction_applied": true,
            "evaluator_type": res.evaluator_type,
            "effects": response.effects,
        });
        telemetry.emit_async(event, dek_telemetry::spooler::Priority::Normal);
    }

    (StatusCode::OK, Json(json_resp))
}

async fn batch_check(
    State(state): State<ApiState>,
    Json(reqs): Json<Vec<DecisionRequest>>,
) -> impl IntoResponse {
    let mut responses = Vec::with_capacity(reqs.len());

    let enf = state.enforcement.load();
    let health = *state.identity_health.borrow();
    let is_strict_deny = matches!(**enf, EnforcementState::StrictDeny { .. })
        || health == crate::svid_renewal_failclosed::IdentityHealth::Expired;

    let deny_reason = if health == crate::svid_renewal_failclosed::IdentityHealth::Expired {
        "fail-closed: SVID expired".to_string()
    } else if let EnforcementState::StrictDeny { ref reason, .. } = **enf {
        format!("fail-closed: {}", reason)
    } else {
        String::new()
    };

    // Simple serial evaluation for now. Could be parallelized.
    for req in reqs {
        if is_strict_deny {
            responses.push(DecisionResponse {
                decision_id: uuid::Uuid::new_v4().to_string(),
                allow: false,
                reason_code: "DENIED_BY_FAIL_CLOSED".into(),
                reason: deny_reason.clone(),
                obligations: vec![],
                effects: serde_json::json!({}),
                policy_bundle_id: "active".into(),
                policy_bundle_version: "v1".into(),
                evaluator_results: vec![],
                latency_ms: 0_i64,
            });
            continue;
        }

        let snap = state.snapshot.load();
        let val = serde_json::to_value(&req).unwrap_or(serde_json::json!({}));

        let start = std::time::Instant::now();
        let res = snap.router.authorize(val).await.unwrap_or_else(|_| {
            dek_policy_runtime::PolicyDecision {
                evaluator_id: "core_api".into(),
                evaluator_type: "router".into(),
                required: true,
                status: "error".into(),
                decision: "deny".into(),
                allow: false,
                reason: "Policy evaluation failed".into(),
                effects: serde_json::json!({}),
                obligations: vec![],
                metadata: serde_json::json!({}),
                explanation: None,
                user_action_required: false,
                user_action_th: None,
            }
        });
        let latency = start.elapsed().as_millis() as i64;

        let response = DecisionResponse {
            decision_id: uuid::Uuid::new_v4().to_string(),
            allow: res.allow,
            reason_code: if res.allow {
                "OK".into()
            } else {
                "DENIED_BY_POLICY".into()
            },
            reason: res.reason.clone(),
            obligations: vec![],
            effects: res.effects.clone(),
            policy_bundle_id: "active".into(),
            policy_bundle_version: "v1".into(),
            evaluator_results: vec![],
            latency_ms: latency,
        };

        if let Some(telemetry) = &state.telemetry {
            let event = serde_json::json!({
                "schema_version": "1.0",
                "event_id": response.decision_id.clone(),
                "event_type": "decision",
                "timestamp": chrono::Utc::now().to_rfc3339(),
                "tenant_id": req.context.get("tenant_id").and_then(|v| v.as_str()).unwrap_or("local").to_string(),
                "device_id": "api-sidecar",
                "trace_id": "",
                "span_id": "",
                "spiffe_id": "",
                "pep_type": "api",
                "agent_id": "dek",
                "principal_id": req.principal.clone(),
                "mcp_server_id": "",
                "tool_id": "",
                "tool_name": "",
                "action": req.action.clone(),
                "resource_id": req.resource.clone(),
                "resource_uri": "",
                "decision": if res.allow { "allow" } else { "deny" },
                "reason": res.reason.clone(),
                "policy_ids": Vec::<String>::new(),
                "bundle_id": "active",
                "bundle_version": "v1",
                "latency_ms": latency,
                "cached": false,
                "redaction_applied": true,
                "evaluator_type": res.evaluator_type.clone(),
                "effects": response.effects.clone(),
            });
            telemetry.emit_async(event, dek_telemetry::spooler::Priority::Normal);
        }

        responses.push(response);
    }

    (StatusCode::OK, Json(responses))
}

async fn a2a_message(
    State(state): State<ApiState>,
    Json(msg): Json<dek_a2a_mediator::A2AMessage>,
) -> impl IntoResponse {
    let broker = dek_a2a_mediator::IATPBroker::new();

    let sender_trust = dek_agent_observer::trust::TrustScore {
        agent_id: msg.sender_id.clone(),
        score: 1.0,
        reasons: vec![],
    };

    let result = broker.route_message(&msg, &sender_trust);

    // --- Emit ToolUsagePayload ---
    if let Some(telemetry) = &state.telemetry {
        let decision_outcome = match &result {
            Ok(_) => "allow",
            Err(_) => "deny",
        };
        let tool_payload = serde_json::json!({
            "agent_id": msg.sender_id.clone(),
            "agent_label": msg.sender_id.clone(),
            "tool_kind": "a2a_skill",
            "tool_name": format!("a2a.send_to_{}", msg.receiver_id),
            "server": "a2a-mediator",
            "decision": decision_outcome,
            "enforced_for_real": true,
            "args_redacted": "<redacted>",
            "observed_at": chrono::Utc::now().to_rfc3339()
        });
        let env = serde_json::json!({
            "schema_version": "telemetry-envelope.v1",
            "event_id": uuid::Uuid::new_v4().to_string(),
            "event_type": "tool_usage",
            "timestamp": chrono::Utc::now().to_rfc3339(),
            "tenant_id": "local",
            "device_id": "dek-local",
            "redaction_applied": false,
            "payload": tool_payload
        });
        telemetry.emit_async(env, dek_telemetry::spooler::Priority::Normal);
    }

    match result {
        Ok(_) => (
            StatusCode::OK,
            Json(serde_json::json!({ "status": "delivered" })),
        ),
        Err(e) => {
            tracing::warn!("A2A Message Blocked: {}", e);
            (
                StatusCode::FORBIDDEN,
                Json(serde_json::json!({ "status": "blocked", "reason": e.to_string() })),
            )
        }
    }
}